Initialise OpenSSL before calling chroot()
OpenSSL may need to load additional shared libraries, in particular
for the gost algorithm.  This will not work after we chroot(), so we
need to initialise it before doing that.

Move the calls to dst_lib_init2() and isc_entropy_create() into
setup() and calls to the corresponding cleanup into cleanup().

Addresses-Debian-Bug: 696661
Signed-off-by: LaMont Jones <lamont@debian.org>